requirements.txt
python-dotenv
PyJWT
flask
flask_cors
flask_pymongo
web3
eth-account
bson